Homework_14/requirements.txt
python-dotenv~=1.0.1
pydantic~=2.6.3
Faker~=22.2.0
anyio~=4.2.0
fastapi~=0.110.0
SQLAlchemy~=2.0.28
passlib~=1.7.4
alembic~=1.13.1
libgravatar~=1.0.4
jose~=1.0.0
aiohttp~=3.9.3
uvicorn~=0.28.0
pytest~=8.1.1
redis~=5.0.3
cloudinary~=1.39.0
colorlog~=6.8.2